+++ server/finatra-server/src/main/scala/tapir/server/finatra/FinatraContent.scala
@@ -1,0 +1,6 @@
+package tapir.server.finatra
+import com.twitter.io.{Buf, Reader}
+
+sealed trait FinatraContent
+case class FinatraContentBuf(buf: Buf) extends FinatraContent
+case class FinatraContentReader(reader: Reader[Buf]) extends FinatraContent
