+++ server/finatra-server/src/main/scala/tapir/server/finatra/FinatraRequestToRawBody.scala
@@ -1,5 +1,5 @@
 package tapir.server.finatra
-import java.io.{ByteArrayInputStream, FileOutputStream}
+import java.io.{ByteArrayInputStream, File, FileOutputStream}
 import java.nio.ByteBuffer
 import java.nio.charset.Charset
 
@@ -38,6 +38,26 @@ object FinatraRequestToRawBody {
       buffer
     }
 
+    rawBodyType match {
+      case StringValueType(defaultCharset) => new String(asByteArray, charset.getOrElse(defaultCharset))
+      case ByteArrayValueType              => asByteArray
+      case ByteBufferValueType             => asByteBuffer
+      case InputStreamValueType            => new ByteArrayInputStream(asByteArray)
+      case FileValueType                   => byteArrayToFile(asByteArray)
+      case mvt: MultipartValueType         => multiPartRequestToRawBody(request, mvt)
+    }
+  }
+
+  private def byteArrayToFile(byteArray: Array[Byte]): File = {
+    // TODO: Make this async
+    val file = Defaults.createTempFile()
+    val outputStream = new FileOutputStream(file)
+    outputStream.write(byteArray)
+    outputStream.close()
+    file
+  }
+
+  private def multiPartRequestToRawBody(request: Request, mvt: MultipartValueType): Seq[RawPart] = {
     def fileItemHeadersToSeq(headers: FileItemHeaders): Seq[(String, String)] = {
       headers.getHeaderNames.asScala
         .flatMap { name =>
@@ -47,53 +67,39 @@ object FinatraRequestToRawBody {
         .filter(_._1.toLowerCase != "content-disposition")
     }
 
-    rawBodyType match {
-      case StringValueType(defaultCharset) => new String(asByteArray, charset.getOrElse(defaultCharset))
-      case ByteArrayValueType              => asByteArray
-      case ByteBufferValueType             => asByteBuffer
-      case InputStreamValueType            => new ByteArrayInputStream(asByteArray)
-      case FileValueType                   =>
-        // TODO: Make this async
-        val file = Defaults.createTempFile()
-        val outputStream = new FileOutputStream(file)
-        outputStream.write(asByteArray)
-        outputStream.close()
-        file
-      case mvt: MultipartValueType =>
-        RequestUtils
-          .multiParams(request)
-          .flatMap {
-            case (name, multiPartItem) =>
-              val dispositionParams = Option(multiPartItem.headers.getHeader("content-disposition"))
-                .map(
-                  _.split(";")
-                    .map(_.trim)
-                    .tail
-                    .map(_.split("="))
-                    .map(array => array(0) -> array(1))
-                    .toMap
-                )
-                .getOrElse(Map.empty)
+    RequestUtils
+      .multiParams(request)
+      .flatMap {
+        case (name, multiPartItem) =>
+          val dispositionParams = Option(multiPartItem.headers.getHeader("content-disposition"))
+            .map(
+              _.split(";")
+                .map(_.trim)
+                .tail
+                .map(_.split("="))
+                .map(array => array(0) -> array(1))
+                .toMap
+            )
+            .getOrElse(Map.empty)
 
-              val charset = multiPartItem.contentType.flatMap(
-                _.split(";")
-                  .map(_.trim)
-                  .tail
-                  .map(_.split("="))
-                  .map(array => array(0) -> array(1))
-                  .toMap
-                  .get("charset")
-                  .map(Charset.forName)
-              )
+          val charset = multiPartItem.contentType.flatMap(
+            _.split(";")
+              .map(_.trim)
+              .tail
+              .map(_.split("="))
+              .map(array => array(0) -> array(1))
+              .toMap
+              .get("charset")
+              .map(Charset.forName)
+          )
 
-              mvt
-                .partCodecMeta(name)
-                .map(codecMeta => apply(codecMeta.rawValueType, Buf.ByteArray.Owned(multiPartItem.data), charset, request))
-                .map(
-                  body => Part(name, dispositionParams - "name", fileItemHeadersToSeq(multiPartItem.headers), body).asInstanceOf[RawPart]
-                )
-          }
-          .toSeq
-    }
+          mvt
+            .partCodecMeta(name)
+            .map(codecMeta => apply(codecMeta.rawValueType, Buf.ByteArray.Owned(multiPartItem.data), charset, request))
+            .map(
+              body => Part(name, dispositionParams - "name", fileItemHeadersToSeq(multiPartItem.headers), body).asInstanceOf[RawPart]
+            )
+      }
+      .toSeq
   }
 }
