+++ server/finatra-server/src/main/scala/tapir/server/finatra/OutputToFinatraResponse.scala
@@ -2,7 +2,7 @@ package tapir.server.finatra
 import java.io.{File, InputStream}
 import java.nio.ByteBuffer
 
-import com.twitter.finagle.http.{Response, Status, Version}
+import com.twitter.finagle.http.Status
 import com.twitter.io.{Buf, InputStreamReader, Reader}
 import org.apache.http.entity.ContentType
 import org.apache.http.entity.mime.content._
@@ -22,42 +22,6 @@ import tapir.{
   RawPart,
   StreamingEndpointIO,
   StringValueType
-}
-
-sealed trait FinatraContent
-case class FinatraContentBuf(buf: Buf) extends FinatraContent
-case class FinatraContentReader(reader: Reader[Buf]) extends FinatraContent
-
-case class FinatraResponse(
-    status: Status,
-    content: FinatraContent = FinatraContentBuf(Buf.Empty),
-    contentType: String = "text/plain",
-    headerMap: Seq[(String, String)] = Seq.empty
-) {
-  def toResponse: Response = {
-    val responseWithContent = content match {
-      case FinatraContentBuf(buf) =>
-        val response = Response(Version.Http11, status)
-        response.content = buf
-        response
-      case FinatraContentReader(reader) =>
-        Response(Version.Http11, status, reader)
-    }
-    responseWithContent.contentType = contentType
-    headerMap.foreach { case (name, value) => responseWithContent.headerMap.add(name, value) }
-
-    responseWithContent
-  }
-
-  // There should only be one content-type header, so if we're
-  // adding a content-type, use 'set' rather than 'add'.
-  def setOrAddHeader(name: String, value: String): FinatraResponse = {
-    if (name.toLowerCase() == "content-type") {
-      this.copy(contentType = value)
-    } else {
-      this.copy(headerMap = this.headerMap :+ (name -> value))
-    }
-  }
 }
 
 object OutputToFinatraResponse {
